Update dependencies.  Move package.js so types are more sane.
