+++ lib/index.js
@@ -5,7 +5,7 @@ import {dirname} from 'node:path';
 import {findRoot} from './findRoot.js';
 import {parse} from 'semver';
 import {stringify} from './stringify.js';
-import {version} from '../package.js';
+import {version} from './package.js';
 
 const PREAMBLE = `\
 // Generated by package-extract v${version}
