+++ package.json
@@ -25,9 +25,9 @@
   "scripts": {
     "lint": "eslint .",
     "test": "c8 node --test",
-    "ver": "node ./bin/package-extract.js -s",
+    "ver": "node ./bin/package-extract.js -s -o lib/package.js",
     "build": "npm run ver && npm run types && npm test && npm run lint && npm pack --dry-run",
-    "types": "tsc",
+    "types": "tsgo",
     "version": "npm run ver && npm run build && git add package.js types/package.d.ts"
   },
   "keywords": [
@@ -43,19 +43,19 @@
   "author": "Joe Hildebrand <joe-github@cursive.net>",
   "license": "MIT",
   "dependencies": {
-    "commander": "~14.0.3",
-    "semver": "~7.7.4"
+    "commander": "~15.0.0",
+    "semver": "~7.8.2"
   },
   "devDependencies": {
-    "@cto.af/eslint-config": "^7.0.0",
-    "@eslint/markdown": "^8.0.1",
-    "@types/node": "^25.6.0",
+    "@cto.af/eslint-config": "^7.0.2",
+    "@eslint/markdown": "^8.0.2",
+    "@typescript/native-preview": "7.0.0-dev.20260606.1",
+    "@types/node": "^25.9.2",
     "@types/semver": "^7.7.1",
-    "eslint": "^10.2.1",
-    "superc8": "^12.3.1",
-    "typescript": "^6.0.3"
+    "eslint": "^10.4.1",
+    "superc8": "^12.7.0"
   },
-  "packageManager": "pnpm@11.0.1",
+  "packageManager": "pnpm@11.5.2",
   "engines": {
     "node": ">=22"
   }
