+++ pnpm-workspace.yaml
@@ -1,0 +1,11 @@
+minimumReleaseAgeExclude:
+  - '@cto.af/eslint-config@7.0.2'
+  - '@types/node@25.9.2'
+  - '@typescript/native-preview-darwin-arm64@7.0.0-dev.20260606.1'
+  - '@typescript/native-preview-darwin-x64@7.0.0-dev.20260606.1'
+  - '@typescript/native-preview-linux-arm64@7.0.0-dev.20260606.1'
+  - '@typescript/native-preview-linux-arm@7.0.0-dev.20260606.1'
+  - '@typescript/native-preview-linux-x64@7.0.0-dev.20260606.1'
+  - '@typescript/native-preview-win32-arm64@7.0.0-dev.20260606.1'
+  - '@typescript/native-preview-win32-x64@7.0.0-dev.20260606.1'
+  - '@typescript/native-preview@7.0.0-dev.20260606.1'
